Moved data recording control to a separate component
